pyproject.toml
[tool.poetry]
name = "py_instagram_unfollowers"
version = "0.1.0"
description = "Find out who is not following you back on Instagram."
authors = ["Kevin Rodríguez <kyrodriguez97@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
selenium = "^4.7.2"
webdriver-manager = "^3.8.5"
redis = "^4.6.0"
sendgrid = "^6.10.0"

[tool.poetry.dev-dependencies]
isort = "^5.11.4"
python-dotenv = "^0.21.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
